pyproject.toml
[tool.poetry]
name = "hr-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.3.0"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.25"
openai = "^1.8.0"
alembic = "^1.13.1"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
